Moved example code that was in App.js to a folder called data-entry. Have Set up Routes in App.js and the BrowserRouter in index.js
